Check patch parameter types.
BUG=
R=karlklose@google.com

Review URL: https://codereview.chromium.org//15022005

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge/dart@22729 260f80e4-7a28-3924-810f-c04153c831b5

+++ sdk/lib/_internal/compiler/implementation/lib/js_helper.dart
@@ -269,7 +269,7 @@ class Primitives {
     return JS('num', r'parseInt(#, #)', source, radix);
   }
 
-  static double parseDouble(String source, int handleError(String source)) {
+  static double parseDouble(String source, double handleError(String source)) {
     checkString(source);
     if (handleError == null) handleError = _throwFormatException;
     // Notice that JS parseFloat accepts garbage at the end of the string.
